+++ backend/src/main/java/com/ice/studyroom/domain/membership/presentation/MembershipController.java
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.ice.studyroom.domain.identity.domain.JwtToken;
+import com.ice.studyroom.global.security.jwt.JwtToken;
 import com.ice.studyroom.domain.membership.application.MembershipService;
 import com.ice.studyroom.domain.membership.presentation.dto.request.EmailVerificationRequest;
 import com.ice.studyroom.domain.membership.presentation.dto.request.MemberCreateRequest;
