+++ backend/src/main/java/com/ice/studyroom/domain/reservation/application/ReservationService.java
@@ -19,7 +19,7 @@ import org.springframework.util.ObjectUtils;
 
 import com.ice.studyroom.domain.admin.domain.type.RoomType;
 import com.ice.studyroom.domain.reservation.infrastructure.redis.QRCodeService;
-import com.ice.studyroom.domain.identity.domain.service.TokenService;
+import com.ice.studyroom.global.security.service.TokenService;
 import com.ice.studyroom.domain.reservation.infrastructure.util.QRCodeUtil;
 import com.ice.studyroom.domain.membership.domain.entity.Member;
 import com.ice.studyroom.domain.membership.domain.service.MemberDomainService;
